updated writing data script
